restructure dmatrix
