+++ src/blaslapack/mod.rs
@@ -17,16 +17,16 @@
 //}}}
 //--------------------------------------------------------------------------------------------------
 
-pub mod common;
-pub mod gees;
-pub mod geev;
-pub mod gemm;
-pub mod gemv;
-pub mod geqrf;
-pub mod gesv;
-pub mod getrf;
-pub mod orgqr;
-pub mod syev;
+mod common;
+mod gees;
+mod geev;
+mod gemm;
+mod gemv;
+mod geqrf;
+mod gesv;
+mod getrf;
+mod orgqr;
+mod syev;
 
 /// Abstracts over matrix storage for generic LAPACK dispatch.
 #[allow(dead_code)]
@@ -36,3 +36,15 @@ pub(crate) trait MatrixBuffer: crate::common::Shape
     fn as_slice(&self) -> &[Self::Scalar];
     fn as_mut_slice(&mut self) -> &mut [Self::Scalar];
 }
+
+
+pub(crate) use common::AsI32;
+pub(crate) use gees::{Error as ShurRawError, schur_raw, Gees};
+pub(crate) use geev::{Error as EigRawError, eig_raw, Geev};
+pub(crate) use gemm::{matmul_dispatch, Gemm};
+pub(crate) use gemv::{Gemv};
+pub(crate) use geqrf::{QrRawError, qr_raw, Geqrf};
+pub(crate) use gesv::{Error as SolveRawError, solve_raw, Gesv};
+pub(crate) use getrf::{Error as LuRawError, lu_raw, Getrf};
+pub(crate) use orgqr::{Orgqr};
+pub(crate) use syev::{Error as SymEigRawError, symeig_raw, Syev};
