+++ src/dmatrix/blaslapack/matmul.rs
@@ -8,10 +8,8 @@
 //--------------------------------------------------------------------------------------------------
 
 //{{{ crate imports
-use super::DMatrix;
-use crate::blaslapack::gemm::matmul_dispatch;
-use crate::blaslapack::gemm::Gemm;
-use crate::blaslapack::gemv::Gemv;
+use crate::dmatrix::DMatrix;
+use crate::blaslapack::{Gemm, Gemv, matmul_dispatch};
 use crate::common::{Field, MatMul, One, Zero};
 use crate::smatrix::SMatrix;
 //}}}
