+++ src/dmatrix/blaslapack/mod.rs
@@ -1,0 +1,21 @@
+//! Short Description of module
+//!
+//! Longer description of module
+//--------------------------------------------------------------------------------------------------
+
+//{{{ crate imports
+//}}}
+//{{{ std imports
+//}}}
+//{{{ dep imports
+//}}}
+//--------------------------------------------------------------------------------------------------
+
+
+pub mod eig;
+pub mod lu;
+pub mod matmul;
+pub mod qr;
+pub mod schur;
+pub mod solve;
+pub mod symeig;
