+++ src/dmatrix/elementwise/addop.rs
@@ -1,54 +1,57 @@
-//! Element-wise multiplication operators for [`DMatrix`]: matrix * scalar and matrix * matrix.
+//! Addition operators for [`DMatrix`]: matrix + matrix and matrix + scalar.
 //!
-//! Implements the [`Mul`] trait for element-wise (Hadamard) multiplication of [`DMatrix<T>`]
-//! operands and for scalar–matrix scaling. This is *not* matrix multiplication; for that see
-//! `matmul`. Both matrix–matrix and scalar–matrix products are lazy, returning a
-//! `BinopExpr` that is evaluated on demand when converted into a concrete [`DMatrix`].
+//! Implements the standard [`Add`] trait for all combinations of owned and borrowed
+//! [`DMatrix<T>`] operands, as well as mixed matrix–scalar and scalar–matrix overloads. For
+//! matrix–matrix addition the result is a lazy `BinopExpr` that defers allocation until the
+//! expression is converted into a concrete [`DMatrix`]; for scalar operands the addition is
+//! applied element-wise.
+//!
 //--------------------------------------------------------------------------------------------------
 
 //{{{ crate imports
-use super::DMatrix;
+use crate::dmatrix::DMatrix;
 use crate::apply_for_all_types;
 #[cfg(feature = "enable_checks")]
 use crate::common::Shape;
 use crate::common::{Field, IndexValue, LazyExpr};
-use crate::expression::binary_expr::{BinopExpr, MulOp};
+use crate::expression::binary_expr::{AddOp, BinopExpr};
 //}}}
 //{{{ std imports
-use std::ops::{Mul, MulAssign};
+use std::ops::{Add, AddAssign};
 //}}}
 //{{{ dep imports
 //}}}
 //--------------------------------------------------------------------------------------------------
+
 //{{{ collection: eagerly evaluated expressions
-//{{{ impl: Mul<T> for DMatrix
-impl<T> Mul<T> for DMatrix<T>
+//{{{ impl: Add<T> for DMatrix
+impl<T> Add<T> for DMatrix<T>
 where
     T: Field + Copy,
 {
     type Output = DMatrix<T>;
 
     #[inline]
-    fn mul(
+    fn add(
         self,
         rhs: T,
     ) -> Self::Output
     {
         let mut out = self.clone();
-        out.iter_mut().for_each(|x| *x *= rhs);
+        out.iter_mut().for_each(|x| *x += rhs);
         out
     }
 }
 //}}}
-//{{{ impl: Mul<DMatrix> for DMatrix
-impl<T> Mul for DMatrix<T>
+//{{{ impl: Add<DMatrix> for DMatrix
+impl<T> Add for DMatrix<T>
 where
     T: Field + Copy + IndexValue<usize, Output = T>,
 {
     type Output = DMatrix<T>;
 
     #[inline]
-    fn mul(
+    fn add(
         self,
         rhs: DMatrix<T>,
     ) -> Self::Output
@@ -64,58 +67,58 @@ where
         out.iter_mut()
             .zip(rhs.iter())
             .for_each(|(out_elem, rhs_elem)| {
-                *out_elem *= *rhs_elem;
+                *out_elem += *rhs_elem;
             });
 
         out
     }
 }
 //}}}
-//{{{ impl Mul<DMatrix<T>> for T
-macro_rules! impl_dmatrix_scalar_mul {
+//{{{ impl Add<DMatrix<T>> for T
+macro_rules! impl_dmatrix_add {
     ($type: ty) => {
         #[doc(hidden)]
-        impl Mul<DMatrix<$type>> for $type
+        impl Add<DMatrix<$type>> for $type
         {
             type Output = DMatrix<$type>;
 
             #[inline]
-            fn mul(
+            fn add(
                 self,
                 rhs: DMatrix<$type>,
             ) -> Self::Output
             {
                 let mut out = rhs.clone();
-                out.iter_mut().for_each(|x| *x *= self);
+                out.iter_mut().for_each(|x| *x += self);
                 out
             }
         }
     };
 }
-apply_for_all_types!(impl_dmatrix_scalar_mul);
+apply_for_all_types!(impl_dmatrix_add);
 //}}}
-//{{{ impl MulAssign<T> for DMatrix
-impl<T> MulAssign<T> for DMatrix<T>
+//{{{ impl AddAssign<T> for DMatrix
+impl<T> AddAssign<T> for DMatrix<T>
 where
     T: Field + Copy,
 {
     #[inline]
-    fn mul_assign(
+    fn add_assign(
         &mut self,
         rhs: T,
     )
     {
-        self.iter_mut().for_each(|x| *x *= rhs);
+        self.iter_mut().for_each(|x| *x += rhs);
     }
 }
 //}}}
-//{{{ impl: MulAssign<DMatrix> for DMatrix
-impl<T> MulAssign for DMatrix<T>
+//{{{ impl: AddAssign<DMatrix> for DMatrix
+impl<T> AddAssign for DMatrix<T>
 where
     T: Field + Copy,
 {
     #[inline]
-    fn mul_assign(
+    fn add_assign(
         &mut self,
         rhs: DMatrix<T>,
     )
@@ -130,23 +133,23 @@ where
         self.iter_mut()
             .zip(rhs.iter())
             .for_each(|(out_elem, rhs_elem)| {
-                *out_elem *= *rhs_elem;
+                *out_elem += *rhs_elem;
             });
     }
 }
 //}}}
 //}}}
-//{{{ collection: MulOp for DMatrix
-//{{{ impl: Mul<T> for DMatrix
-macro_rules! impl_dmatrix_mul_scalar_rhs {
+//{{{ collection: Lazily evaluated expressions
+//{{{ impl: Add<T> for &'a DMatrix
+macro_rules! impl_dmatrix_add_scalar_rhs {
     ($type:ty) => {
         #[doc(hidden)]
-        impl<'a> Mul<$type> for &'a DMatrix<$type>
+        impl<'a> Add<$type> for &'a DMatrix<$type>
         {
-            type Output = BinopExpr<&'a DMatrix<$type>, $type, $type, MulOp>;
+            type Output = BinopExpr<&'a DMatrix<$type>, $type, $type, AddOp>;
 
             #[inline]
-            fn mul(
+            fn add(
                 self,
                 rhs: $type,
             ) -> Self::Output
@@ -164,35 +167,35 @@ macro_rules! impl_dmatrix_mul_scalar_rhs {
         }
 
         #[doc(hidden)]
-        impl<'a> Mul<$type> for &'a mut DMatrix<$type>
+        impl<'a> Add<$type> for &'a mut DMatrix<$type>
         {
-            type Output = BinopExpr<&'a DMatrix<$type>, $type, $type, MulOp>;
+            type Output = BinopExpr<&'a DMatrix<$type>, $type, $type, AddOp>;
 
             #[inline]
-            fn mul(
+            fn add(
                 self,
                 rhs: $type,
             ) -> Self::Output
             {
-                (&*self).mul(rhs)
+                (&*self).add(rhs)
             }
         }
     };
 }
 
-apply_for_all_types!(impl_dmatrix_mul_scalar_rhs);
+apply_for_all_types!(impl_dmatrix_add_scalar_rhs);
 
 //}}}
-//{{{ impl: Mul<DMatrix> for $type
-macro_rules! impl_dmatrix_mul {
+//{{{ impl: Add<Dmatrix> for $type
+macro_rules! impl_dmatrix_ref_add {
     ($type:ty) => {
         #[doc(hidden)]
-        impl<'a> Mul<&'a DMatrix<$type>> for $type
+        impl<'a> Add<&'a DMatrix<$type>> for $type
         {
-            type Output = BinopExpr<$type, &'a DMatrix<$type>, $type, MulOp>;
+            type Output = BinopExpr<$type, &'a DMatrix<$type>, $type, AddOp>;
 
             #[inline]
-            fn mul(
+            fn add(
                 self,
                 rhs: &'a DMatrix<$type>,
             ) -> Self::Output
@@ -210,39 +213,39 @@ macro_rules! impl_dmatrix_mul {
         }
     };
 }
-apply_for_all_types!(impl_dmatrix_mul);
+apply_for_all_types!(impl_dmatrix_ref_add);
 //}}}
-//{{{ impl: Mul<&mut DMatrix> for $type
-macro_rules! impl_dmatrix_ref_mut_mul {
+//{{{ impl: Add<&mut Dmatrix> for $type
+macro_rules! impl_dmatrix_ref_mut_add {
     ($type:ty) => {
         #[doc(hidden)]
-        impl<'a> Mul<&'a mut DMatrix<$type>> for $type
+        impl<'a> Add<&'a mut DMatrix<$type>> for $type
         {
-            type Output = BinopExpr<$type, &'a DMatrix<$type>, $type, MulOp>;
+            type Output = BinopExpr<$type, &'a DMatrix<$type>, $type, AddOp>;
 
             #[inline]
-            fn mul(
+            fn add(
                 self,
                 rhs: &'a mut DMatrix<$type>,
             ) -> Self::Output
             {
-                self.mul(&*rhs)
+                self.add(&*rhs)
             }
         }
     };
 }
-apply_for_all_types!(impl_dmatrix_ref_mut_mul);
+apply_for_all_types!(impl_dmatrix_ref_mut_add);
 //}}}
-//{{{ impl: Mul<Rhs> for &'a DMatrix
-impl<'a, T, Rhs> Mul<Rhs> for &'a DMatrix<T>
+//{{{ impl: Add<Rhs> for &'a DMatrix
+impl<'a, T, Rhs> Add<Rhs> for &'a DMatrix<T>
 where
     T: Field + Copy,
     Rhs: LazyExpr<ScalarType = T> + IndexValue<usize, Output = T>,
 {
-    type Output = BinopExpr<&'a DMatrix<T>, Rhs, T, MulOp>;
+    type Output = BinopExpr<&'a DMatrix<T>, Rhs, T, AddOp>;
 
     #[inline]
-    fn mul(
+    fn add(
         self,
         rhs: Rhs,
     ) -> Self::Output
@@ -266,21 +269,21 @@ where
 }
 
 //}}}
-//{{{ impl: Mul<Rhs> for &'a mut DMatrix
-impl<'a, T, Rhs> Mul<Rhs> for &'a mut DMatrix<T>
+//{{{ impl: Add<Rhs> for &'a mut DMatrix
+impl<'a, T, Rhs> Add<Rhs> for &'a mut DMatrix<T>
 where
     T: Field + Copy,
     Rhs: LazyExpr<ScalarType = T> + IndexValue<usize, Output = T>,
 {
-    type Output = BinopExpr<&'a DMatrix<T>, Rhs, T, MulOp>;
+    type Output = BinopExpr<&'a DMatrix<T>, Rhs, T, AddOp>;
 
     #[inline]
-    fn mul(
+    fn add(
         self,
         rhs: Rhs,
     ) -> Self::Output
     {
-        (&*self).mul(rhs)
+        (&*self).add(rhs)
     }
 }
