+++ src/dmatrix/elementwise/mod.rs
@@ -1,0 +1,18 @@
+//! Short Description of module
+//!
+//! Longer description of module
+//--------------------------------------------------------------------------------------------------
+
+//{{{ crate imports
+//}}}
+//{{{ std imports
+//}}}
+//{{{ dep imports
+//}}}
+//--------------------------------------------------------------------------------------------------
+
+mod addop;
+mod divop;
+mod mulop;
+mod negop;
+mod subop;
