+++ src/dmatrix/elementwise/negop.rs
@@ -8,7 +8,7 @@
 //--------------------------------------------------------------------------------------------------
 
 //{{{ crate imports
-use super::DMatrix;
+use crate::dmatrix::DMatrix;
 use crate::common::{Field, Zero};
 use crate::expression::unary_expr::{NegOp, UnaryExpr};
 //}}}
