+++ src/dmatrix/mod.rs
@@ -23,30 +23,17 @@ use serde::{Deserialize, Serialize};
 //}}}
 //--------------------------------------------------------------------------------------------------
 
-// binary operations
-pub mod addop;
-pub mod divop;
-pub mod mulop;
-pub mod subop;
-// unary operations
-pub mod negop;
-// matrix operations
-pub mod eig;
-pub mod lu;
-pub mod matmul;
-pub mod matrix_ops;
-pub mod qr;
-pub mod reduce_ops;
-pub mod schur;
-pub mod solve;
-pub mod symeig;
-pub mod transform_ops;
-// everything else
-pub mod construction;
-pub mod indexing;
-pub mod io;
-pub mod iteration;
-pub mod subviews;
+// elementwise expressions
+mod elementwise;
+mod blaslapack;
+mod matrix_ops;
+mod reduce_ops;
+mod transform_ops;
+mod construction;
+mod indexing;
+mod io;
+mod iteration;
+mod subviews;
 
 //{{{ collection: DMatrix
 //{{{ struct: DMatrix
@@ -220,7 +207,7 @@ where
         DMatrix { data, nrows, ncols }
     }
 } //}}}
-  //{{{ impl: From<UnaryExpr> for DMatrix
+//{{{ impl: From<UnaryExpr> for DMatrix
 impl<A, T, Op> From<UnaryExpr<A, T, Op>> for DMatrix<T>
 where
     A: IndexValue<usize, Output = T> + crate::common::Shape,
@@ -241,8 +228,9 @@ where
         unsafe { data.set_len(total) };
         DMatrix { data, nrows, ncols }
     }
-} //}}}
-  //}}}
+}
+//}}}
+//}}}
   //{{{ collection: DVector
   //{{{ type: DVector
 /// A dynamic vector stored as a single-row or single-column [`DMatrix`].
